requirements.txt
flask==1.1.2
gunicorn==20.0.4
pytest==5.4.1
jwt==1.0.0
pyjwt==1.7.1
PyYAML==5.3.1
psycopg2-binary==2.8.5
firebase-admin==4.1.0
requests==2.23.0
flask-swagger-ui==3.25.0
flask-cors==3.0.8
sendgrid==6.2.2
pytest-postgresql==2.3.0
Flask-HTTPAuth==4.0.0